add some more features
