--- modules/kb.core.bowline/src/main/kotlin/kb/core/bowline/BowlineTest.kt
@@ -1,13 +1,0 @@
-package kb.core.bowline
-
-import javafx.scene.Scene
-import javafx.stage.Stage
-
-fun testBowline() {
-    val stage = Stage()
-    stage.title = "Bowline Test"
-    stage.scene = Scene(Bowline())
-    stage.width = 900.0
-    stage.height = 600.0
-    stage.show()
-}
